requirements.txt
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.5.3
contourpy==1.0.6
cycler==0.11.0
dash==2.6.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
Flask-Compress==1.13
fonttools==4.38.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.17.0
ipython==8.6.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter-client==7.4.4
jupyter-core==4.11.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nbformat==5.5.0
nest-asyncio==1.5.6
numpy==1.23.4
open3d==0.16.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
plotly==5.11.0
prompt-toolkit==3.0.31
psutil==5.9.3
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
safetensors==0.2.4
six==1.16.0
stack-data==0.6.0
tenacity==8.1.0
tokenizers==0.13.1
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.24.0
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.3
zipp==3.10.0
